move the component to profile directory
